refactor: move createDbConnection to a separate file to avoid side effect
