+++ src/libs/DB.ts
@@ -1,27 +1,11 @@
 import type { NodePgDatabase } from 'drizzle-orm/node-postgres';
-import { drizzle } from 'drizzle-orm/node-postgres';
-import { Pool } from 'pg';
-import * as schema from '@/models/Schema';
+import type * as schema from '@/models/Schema';
+import { createDbConnection } from '@/utils/DBConnection';
 import { Env } from './Env';
 
 // Stores the db connection in the global scope to prevent multiple instances due to hot reloading with Next.js
 const globalForDb = globalThis as unknown as {
   drizzle: NodePgDatabase<typeof schema>;
-};
-
-// Need a database for production? Check out https://www.prisma.io/?via=nextjsboilerplate
-// Tested and compatible with Next.js Boilerplate
-export const createDbConnection = () => {
-  const pool = new Pool({
-    connectionString: Env.DATABASE_URL,
-    ssl: !Env.DATABASE_URL.includes('localhost') && !Env.DATABASE_URL.includes('127.0.0.1'),
-    max: 1,
-  });
-
-  return drizzle({
-    client: pool,
-    schema,
-  });
 };
 
 const db = globalForDb.drizzle || createDbConnection();
