+++ src/utils/DBMigration.ts
@@ -1,6 +1,6 @@
 import path from 'node:path';
 import { migrate } from 'drizzle-orm/node-postgres/migrator';
-import { createDbConnection } from '@/libs/DB';
+import { createDbConnection } from './DBConnection';
 
 // Create a new and dedicated database connection for running migrations
 const db = createDbConnection();
